Use new 'getPositionState()' for 'anchored()' mixin; move to folder
